Move FeedDetectorBolt.java from bootstrap package to news package
