+++ src/main/java/org/commoncrawl/stormcrawler/news/FeedDetectorBolt.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.commoncrawl.stormcrawler.news.bootstrap;
+package org.commoncrawl.stormcrawler.news;
 
 import java.util.Map;
 
@@ -19,7 +19,6 @@ import org.apache.storm.task.OutputCollector;
 import org.apache.storm.task.TopologyContext;
 import org.apache.storm.tuple.Tuple;
 import org.apache.storm.tuple.Values;
-import org.commoncrawl.stormcrawler.news.ContentDetector;
 import org.slf4j.LoggerFactory;
 
 import com.digitalpebble.stormcrawler.Constants;
