+++ src/main/java/org/commoncrawl/stormcrawler/news/NewsSiteMapParserBolt.java
@@ -32,7 +32,6 @@ import org.apache.storm.task.OutputCollector;
 import org.apache.storm.task.TopologyContext;
 import org.apache.storm.tuple.Tuple;
 import org.apache.storm.tuple.Values;
-import org.commoncrawl.stormcrawler.news.bootstrap.FeedDetectorBolt;
 import org.slf4j.LoggerFactory;
 
 import com.digitalpebble.stormcrawler.Constants;
